+++ packages/modal/src/Footer.styles.ts
@@ -1,0 +1,13 @@
+import { css } from '@leafygreen-ui/emotion';
+// import { uiColors } from '@leafygreen-ui/palette';
+
+export const footerStyle = css`
+  position: relative;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  display: flex;
+  justify-content: right;
+  flex-direction: row-reverse;
+  padding: 24px 35px 35px;
+`;
